Make individual image loading more extensible
Each method for loading images now has its own loader class, allowing for easier addition of file formats in the future. The same should at some point also be done for saving images.
