+++ src/Image.cpp
@@ -2,15 +2,10 @@
 // It is published under the BSD 3-Clause License within the LICENSE file.
 
 #include <tev/Image.h>
+#include <tev/imageio/ImageLoader.h>
 #include <tev/ThreadPool.h>
 
-#include <ImfChannelList.h>
-#include <ImfInputFile.h>
-#include <ImfInputPart.h>
-#include <ImfMultiPartInputFile.h>
-#include <ImfStdIO.h>
-
-#include <stb_image.h>
+#include <Iex.h>
 
 #include <algorithm>
 #include <array>
@@ -23,46 +18,6 @@ using namespace filesystem;
 using namespace std;
 
 TEV_NAMESPACE_BEGIN
-
-bool isExrFile(ifstream& f) {
-    // Taken from http://www.openexr.com/ReadingAndWritingImageFiles.pdf
-    char b[4];
-    f.read(b, sizeof(b));
-    if (f.gcount() < 4) {
-        return false;
-    }
-
-    bool result = !!f && b[0] == 0x76 && b[1] == 0x2f && b[2] == 0x31 && b[3] == 0x01;
-    f.seekg(0);
-    return result;
-}
-
-bool isPfmFile(ifstream& f) {
-    char b[2];
-    f.read(b, sizeof(b));
-    if (f.gcount() < 2) {
-        return false;
-    }
-
-    bool result = !!f && b[0] == 'P' && (b[1] == 'F' || b[1] == 'f');
-    f.seekg(0);
-    return result;
-}
-
-vector<Channel> makeNChannels(int numChannels, Vector2i size) {
-    vector<Channel> channels;
-    if (numChannels > 1) {
-        const vector<string> channelNames = {"R", "G", "B", "A"};
-        for (int c = 0; c < numChannels; ++c) {
-            string name = c < (int)channelNames.size() ? channelNames[c] : to_string(c);
-            channels.emplace_back(name, size);
-        }
-    } else {
-        channels.emplace_back("L", size);
-    }
-
-    return channels;
-}
 
 atomic<int> Image::sId(0);
 
@@ -82,15 +37,15 @@ Image::Image(const filesystem::path& path, const string& channelSelector)
     }
 
     std::string loadMethod;
-    if (isExrFile(f)) {
-        loadMethod = "OpenEXR";
-        readExr(f);
-    } else if (isPfmFile(f)) {
-        loadMethod = "PFM";
-        readPfm(f);
-    } else {
-        loadMethod = "STBI";
-        readStbi(f);
+    for (const auto& imageLoader : ImageLoader::getLoaders()) {
+        if (imageLoader->canLoadFile(f) || imageLoader == ImageLoader::getLoaders().back()) {
+            loadMethod = imageLoader->name();
+            mData = imageLoader->load(f, mPath, mChannelSelector);
+            break;
+        }
+
+        // Ensure there is no lingering non-zero file cursor.
+        f.seekg(0);
     }
 
     auto end = chrono::system_clock::now();
@@ -160,13 +115,13 @@ vector<string> Image::channelsInLayer(string layerName) const {
     vector<string> result;
 
     if (layerName.empty()) {
-        for (const auto& kv : mChannels) {
+        for (const auto& kv : mData.channels) {
             if (kv.first.find(".") == string::npos) {
                 result.emplace_back(kv.first);
             }
         }
     } else {
-        for (const auto& kv : mChannels) {
+        for (const auto& kv : mData.channels) {
             // If the layer name starts at the beginning, and
             // if no other dot is found after the end of the layer name,
             // then we have found a channel of this layer.
@@ -185,7 +140,7 @@ vector<string> Image::channelsInLayer(string layerName) const {
 string Image::toString() const {
     string result = tfm::format("Path: %s\n\nResolution: (%d, %d)\n\nChannels:\n", mName, size().x(), size().y());
 
-    auto localLayers = mLayers;
+    auto localLayers = mData.layers;
     transform(begin(localLayers), end(localLayers), begin(localLayers), [this](string layer) {
         auto channels = channelsInLayer(layer);
         transform(begin(channels), end(channels), begin(channels), [](string channel) {
@@ -200,305 +155,16 @@ string Image::toString() const {
     return result + join(localLayers, "\n");
 }
 
-void Image::readPfm(ifstream& f) {
-    ThreadPool threadPool;
-
-    string magic;
-    Vector2i size;
-    float scale;
-
-    f >> magic >> size.x() >> size.y() >> scale;
-
-    if (magic != "PF" && magic != "Pf") {
-        throw invalid_argument{tfm::format("Invalid magic PFM string %s in file %s", magic, mPath)};
-    }
-
-    if (!isfinite(scale) || scale == 0) {
-        throw invalid_argument{tfm::format("Invalid PFM scale %f in file %s", scale, mPath)};
-    }
-
-    int numChannels = magic[1] == 'F' ? 3 : 1;
-    bool isPfmLittleEndian = scale < 0;
-    scale = abs(scale);
-
-    vector<Channel> channels = makeNChannels(numChannels, size);
-
-    auto numPixels = (DenseIndex)size.x() * size.y();
-    auto numFloats = numPixels * numChannels;
-    auto numBytes = numFloats * sizeof(float);
-
-    // Skip last newline at the end of the header.
-    string line;
-    getline(f, line);
-
-    // Read entire file in binary mode.
-    vector<float> data(numFloats);
-    f.read(reinterpret_cast<char*>(data.data()), numBytes);
-    if (f.gcount() < (streamsize)numBytes) {
-        throw invalid_argument{tfm::format("Not sufficient bytes to read (%d vs %d) in file %s", f.gcount(), numBytes, mPath)};
-    }
-
-    // Reverse bytes of every float if endianness does not match up with system
-    const bool shallSwapBytes = isSystemLittleEndian() != isPfmLittleEndian;
-
-    threadPool.parallelFor<DenseIndex>(0, size.y(), [&](DenseIndex y) {
-        for (int x = 0; x < size.x(); ++x) {
-            int baseIdx = (y * size.x() + x) * numChannels;
-            for (int c = 0; c < numChannels; ++c) {
-                float val = data[baseIdx + c];
-
-                // Thankfully, due to branch prediction, the "if" in the
-                // inner loop is no significant overhead.
-                if (shallSwapBytes) {
-                    val = swapBytes(val);
-                }
-
-                // Flip image vertically due to PFM format
-                channels[c].at({x, size.y() - y - 1}) = scale * val;
-            }
-        }
-    });
-
-    for (auto& channel : channels) {
-        string name = channel.name();
-        if (matches(name, mChannelSelector, false)) {
-            mChannels.emplace(move(name), move(channel));
-        }
-    }
-
-    // PFM can not contain layers, so all channels simply reside
-    // within a topmost root layer.
-    mLayers.emplace_back("");
-}
-
-void Image::readStbi(ifstream& f) {
-    ThreadPool threadPool;
-
-    static const stbi_io_callbacks callbacks = {
-        // Read
-        [](void* context, char* data, int size) {
-            auto stream = reinterpret_cast<ifstream*>(context);
-            stream->read(data, size);
-            return (int)stream->gcount();
-        },
-        // Seek
-        [](void* context, int size) {
-            reinterpret_cast<ifstream*>(context)->seekg(size, ios_base::cur);
-        },
-        // EOF
-        [](void* context) {
-            return (int)!!(*reinterpret_cast<ifstream*>(context));
-        },
-    };
-
-    void* data;
-    int numChannels;
-    Vector2i size;
-    bool isHdr = stbi_is_hdr_from_callbacks(&callbacks, &f) != 0;
-    f.seekg(0);
-
-    if (isHdr) {
-        data = stbi_loadf_from_callbacks(&callbacks, &f, &size.x(), &size.y(), &numChannels, 0);
-    } else {
-        data = stbi_load_from_callbacks(&callbacks, &f, &size.x(), &size.y(), &numChannels, 0);
-    }
-
-    if (!data) {
-        throw invalid_argument{tfm::format("Could not load texture data from file %s", mPath)};
-    }
-
-    ScopeGuard dataGuard{[data] { stbi_image_free(data); }};
-
-    vector<Channel> channels = makeNChannels(numChannels, size);
-
-    auto numPixels = (DenseIndex)size.x() * size.y();
-    if (isHdr) {
-        auto typedData = reinterpret_cast<float*>(data);
-        threadPool.parallelFor<DenseIndex>(0, numPixels, [&](DenseIndex i) {
-            int baseIdx = i * numChannels;
-            for (int c = 0; c < numChannels; ++c) {
-                channels[c].at(i) = typedData[baseIdx + c];
-            }
-        });
-    } else {
-        auto typedData = reinterpret_cast<unsigned char*>(data);
-        threadPool.parallelFor<DenseIndex>(0, numPixels, [&](DenseIndex i) {
-            int baseIdx = i * numChannels;
-            for (int c = 0; c < numChannels; ++c) {
-                channels[c].at(i) = toLinear((typedData[baseIdx + c]) / 255.0f);
-            }
-        });
-    }
-
-    for (auto& channel : channels) {
-        string name = channel.name();
-        if (matches(name, mChannelSelector, false)) {
-            mChannels.emplace(move(name), move(channel));
-        }
-    }
-
-    // STBI can not load layers, so all channels simply reside
-    // within a topmost root layer.
-    mLayers.emplace_back("");
-}
-
-void Image::readExr(ifstream& f) {
-    ThreadPool threadPool;
-
-    Imf::StdIFStream imfStream{f, mPath.str().c_str()};
-    Imf::MultiPartInputFile multiPartFile{imfStream};
-    int numParts = multiPartFile.parts();
-
-    if (numParts <= 0) {
-        throw invalid_argument{tfm::format("EXR image '%s' does not contain any parts.", mPath)};
-    }
-
-    // Find the first part containing a channel that matches the given channelSubstr.
-    int partIdx = 0;
-    for (int i = 0; i < numParts; ++i) {
-        Imf::InputPart part{multiPartFile, i};
-
-        const Imf::ChannelList& imfChannels = part.header().channels();
-
-        for (Imf::ChannelList::ConstIterator c = imfChannels.begin(); c != imfChannels.end(); ++c) {
-            if (matches(c.name(), mChannelSelector, false)) {
-                partIdx = i;
-                goto l_foundPart;
-            }
-        }
-    }
-l_foundPart:
-
-    Imf::InputPart file{multiPartFile, partIdx};
-    Imath::Box2i dw = file.header().dataWindow();
-    Vector2i size = {dw.max.x - dw.min.x + 1 , dw.max.y - dw.min.y + 1};
-
-    // Inline helper class for dealing with the raw channels loaded from an exr file.
-    class RawChannel {
-    public:
-        RawChannel(string name, Imf::Channel imfChannel)
-        : mName(name), mImfChannel(imfChannel) {
-        }
-
-        void resize(size_t size) {
-            mData.resize(size * bytesPerPixel());
-        }
-
-        void registerWith(Imf::FrameBuffer& frameBuffer, const Imath::Box2i& dw) {
-            int width = dw.max.x - dw.min.x + 1;
-            frameBuffer.insert(mName.c_str(), Imf::Slice(
-                mImfChannel.type,
-                mData.data() - (dw.min.x + dw.min.y * width) * bytesPerPixel(),
-                bytesPerPixel(), bytesPerPixel() * width,
-                mImfChannel.xSampling, mImfChannel.ySampling, 0
-            ));
-        }
-
-        void copyTo(Channel& channel, ThreadPool& threadPool) const {
-            // TODO: Switch to generic lambda once C++14 is used
-            switch (mImfChannel.type) {
-                case Imf::HALF: {
-                    auto data = reinterpret_cast<const half*>(mData.data());
-                    threadPool.parallelForNoWait<DenseIndex>(0, channel.count(), [&, data](DenseIndex i) {
-                        channel.at(i) = data[i];
-                    });
-                    break;
-                }
-
-                case Imf::FLOAT: {
-                    auto data = reinterpret_cast<const float*>(mData.data());
-                    threadPool.parallelForNoWait<DenseIndex>(0, channel.count(), [&, data](DenseIndex i) {
-                        channel.at(i) = data[i];
-                    });
-                    break;
-                }
-
-                case Imf::UINT: {
-                    auto data = reinterpret_cast<const uint32_t*>(mData.data());
-                    threadPool.parallelForNoWait<DenseIndex>(0, channel.count(), [&, data](DenseIndex i) {
-                        channel.at(i) = data[i];
-                    });
-                    break;
-                }
-
-                default:
-                    throw runtime_error("Invalid pixel type encountered.");
-            }
-        }
-
-        const string& name() const {
-            return mName;
-        }
-
-    private:
-        int bytesPerPixel() const {
-            switch (mImfChannel.type) {
-                case Imf::HALF:  return sizeof(half);
-                case Imf::FLOAT: return sizeof(float);
-                case Imf::UINT:  return sizeof(uint32_t);
-                default:
-                    throw runtime_error("Invalid pixel type encountered.");
-            }
-        }
-
-        string mName;
-        Imf::Channel mImfChannel;
-        vector<char> mData;
-    };
-
-    vector<RawChannel> rawChannels;
-    Imf::FrameBuffer frameBuffer;
-
-    const Imf::ChannelList& imfChannels = file.header().channels();
-    set<string> layerNames;
-
-    for (Imf::ChannelList::ConstIterator c = imfChannels.begin(); c != imfChannels.end(); ++c) {
-        if (matches(c.name(), mChannelSelector, false)) {
-            rawChannels.emplace_back(c.name(), c.channel().type);
-            layerNames.insert(Channel::head(c.name()));
-        }
-    }
-
-    if (rawChannels.empty()) {
-        throw invalid_argument{tfm::format("No channels match '%s'.", mChannelSelector)};
-    }
-
-    for (const string& layer : layerNames) {
-        mLayers.emplace_back(layer);
-    }
-
-    threadPool.parallelFor(0, (int)rawChannels.size(), [&](int i) {
-        rawChannels[i].resize((DenseIndex)size.x() * size.y());
-    });
-
-    for (size_t i = 0; i < rawChannels.size(); ++i) {
-        rawChannels[i].registerWith(frameBuffer, dw);
-    }
-
-    file.setFrameBuffer(frameBuffer);
-    file.readPixels(dw.min.y, dw.max.y);
-
-    for (const auto& rawChannel : rawChannels) {
-        mChannels.emplace(rawChannel.name(), Channel{rawChannel.name(), size});
-    }
-
-    for (size_t i = 0; i < rawChannels.size(); ++i) {
-        rawChannels[i].copyTo(mChannels.at(rawChannels[i].name()), threadPool);
-    }
-
-    threadPool.waitUntilFinished();
-}
-
 void Image::ensureValid() {
-    if (mLayers.empty()) {
+    if (mData.layers.empty()) {
         throw runtime_error{"Images must have at least one layer."};
     }
 
-    if (mChannels.empty()) {
+    if (mData.channels.empty()) {
         throw runtime_error{"Images must have at least one channel."};
     }
 
-    for (const auto& kv : mChannels) {
+    for (const auto& kv : mData.channels) {
         if (kv.second.size() != size()) {
             throw runtime_error{tfm::format(
                 "All channels must have the same size as their image. (%s:%dx%d != %dx%d)",
